# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "Exhibition AI API", "version": "0.1.0" },
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "tags": ["auth"],
        "description": "Login with email and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 255,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 1,
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "status",
                        "mustChangePassword",
                        "createdAt"
                      ],
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "email": {
                          "format": "email",
                          "maxLength": 255,
                          "type": "string"
                        },
                        "displayName": {
                          "minLength": 1,
                          "maxLength": 100,
                          "type": "string"
                        },
                        "role": {
                          "anyOf": [
                            { "type": "string", "enum": ["admin"] },
                            { "type": "string", "enum": ["designer"] },
                            { "type": "string", "enum": ["sales"] },
                            { "type": "string", "enum": ["viewer"] }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            { "type": "string", "enum": ["enabled"] },
                            { "type": "string", "enum": ["disabled"] }
                          ]
                        },
                        "mustChangePassword": { "type": "boolean" },
                        "createdAt": { "format": "date-time", "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getMe",
        "tags": ["auth"],
        "description": "Get current user summary.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "status",
                        "mustChangePassword",
                        "createdAt"
                      ],
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "email": {
                          "format": "email",
                          "maxLength": 255,
                          "type": "string"
                        },
                        "displayName": {
                          "minLength": 1,
                          "maxLength": 100,
                          "type": "string"
                        },
                        "role": {
                          "anyOf": [
                            { "type": "string", "enum": ["admin"] },
                            { "type": "string", "enum": ["designer"] },
                            { "type": "string", "enum": ["sales"] },
                            { "type": "string", "enum": ["viewer"] }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            { "type": "string", "enum": ["enabled"] },
                            { "type": "string", "enum": ["disabled"] }
                          ]
                        },
                        "mustChangePassword": { "type": "boolean" },
                        "createdAt": { "format": "date-time", "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/password": {
      "put": {
        "operationId": "changePassword",
        "tags": ["auth"],
        "description": "Change current user password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": ["currentPassword", "newPassword"],
                "properties": {
                  "currentPassword": {
                    "minLength": 1,
                    "maxLength": 255,
                    "type": "string"
                  },
                  "newPassword": {
                    "minLength": 8,
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "No content" },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/send-code": {
      "post": {
        "operationId": "sendVerificationCode",
        "tags": ["auth"],
        "description": "Send an email verification code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": ["email", "type"],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 255,
                    "type": "string"
                  },
                  "type": {
                    "anyOf": [
                      { "type": "string", "enum": ["register"] },
                      { "type": "string", "enum": ["reset_password"] }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "tags": ["auth"],
        "description": "Register a user account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": ["email", "code", "displayName", "password"],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 255,
                    "type": "string"
                  },
                  "code": { "minLength": 6, "maxLength": 6, "type": "string" },
                  "displayName": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 8,
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "status",
                        "mustChangePassword",
                        "createdAt"
                      ],
                      "properties": {
                        "id": { "format": "uuid", "type": "string" },
                        "email": {
                          "format": "email",
                          "maxLength": 255,
                          "type": "string"
                        },
                        "displayName": {
                          "minLength": 1,
                          "maxLength": 100,
                          "type": "string"
                        },
                        "role": {
                          "anyOf": [
                            { "type": "string", "enum": ["admin"] },
                            { "type": "string", "enum": ["designer"] },
                            { "type": "string", "enum": ["sales"] },
                            { "type": "string", "enum": ["viewer"] }
                          ]
                        },
                        "status": {
                          "anyOf": [
                            { "type": "string", "enum": ["enabled"] },
                            { "type": "string", "enum": ["disabled"] }
                          ]
                        },
                        "mustChangePassword": { "type": "boolean" },
                        "createdAt": { "format": "date-time", "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "tags": ["auth"],
        "description": "Reset a password with an email verification code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": ["email", "code", "newPassword"],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 255,
                    "type": "string"
                  },
                  "code": { "minLength": 6, "maxLength": 6, "type": "string" },
                  "newPassword": {
                    "minLength": 8,
                    "maxLength": 255,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": { "204": { "description": "No content" } }
      }
    },
...
  },
...
}